refactor(server): remove `bin` crate Define bin in `api` crate instead
