+++ Cargo.toml
@@ -1,7 +1,6 @@
 [workspace]
 members = [
   "apps/server/api",
-  "apps/server/bin",
   "apps/server/domain",
   "apps/server/infra",
   "apps/server/services",
