+++ apps/client/src/providers/QueryProvider.tsx
@@ -1,12 +1,13 @@
 import { createClient, FetchTransport } from "@rspc/client"
 import { QueryClient } from "@tanstack/react-query"
-import { Procedures } from "l/bindings"
+import type { Procedures } from "l/bindings"
 import { rspc } from "l/rspc"
 import { type JSXElementConstructor, type ReactElement, useState } from "react"
 
 export function QueryProvider({
   children,
 }: {
+  // biome-ignore lint/suspicious/noExplicitAny: necessary to match the rpsc react-query types
   children: ReactElement<any, string | JSXElementConstructor<any>> | undefined
 }) {
   const [queryClient] = useState(() => new QueryClient())
