+++ apps/server/api/Cargo.toml
@@ -6,8 +6,13 @@ edition.workspace = true
 axum.workspace = true
 derive_more.workspace = true
 domain = { path = "../domain" }
+dotenvy.workspace = true
+infra = { path = "../infra" }
 prisma = { path = "../../../prisma" }
 rspc.workspace = true
+rspc-axum.workspace = true
 serde.workspace = true
 services = { path = "../services" }
 specta.workspace = true
+tokio.workspace = true
+tower-http.workspace = true
