--- apps/server/bin/Cargo.toml
@@ -1,14 +1,0 @@
-[package]
-name = "bin"
-edition.workspace = true
-
-[dependencies]
-api = { path = "../api" }
-axum.workspace = true
-dotenvy.workspace = true
-infra = { path = "../infra" }
-prisma = { path = "../../../prisma" }
-rspc-axum.workspace = true
-services = { path = "../services" }
-tokio.workspace = true
-tower-http.workspace = true
